add method in which ids are given
